move actions in ConnectionManager and separate init from normal flow
